Change layout of search bar
